# openapi.yaml
openapi: 3.0.3
info:
  title: PM Tools Integration API
  version: 1.0.0
  description: |
    A single unified REST API that sits in front of multiple project management
    tools (currently **Jira** and **ClickUp**) behind one shared contract. Every
    route is namespaced by `{provider}` (`jira` or `clickup`) so callers use the
    exact same endpoint shape regardless of which underlying tool actually holds
    the ticket.

    ## How to pick the right route (read this first)

    This section exists specifically to help an LLM or automated agent choose
    the correct endpoint without trial and error:

    1. **You almost never need to guess IDs.** If you don't already have a
       `ticketId` or a `projectKeyOrListId`, call `GET /api/{provider}/containers`
       first to discover valid project/list IDs, or ask the user for the ticket
       ID/key they mean.
    2. **"Show me a ticket" / "what does this ticket say" / "get ticket details
       and discussion"** → use `GET /api/{provider}/tickets/{ticketId}`. It
       returns the ticket AND its comments in one call. This is the right
       default for almost every "look up a ticket" request.
    3. **"Get just the ticket fields, I don't care about comments"** (e.g. you
       only need status/assignee/priority and want a smaller, faster response)
       → use `GET /api/{provider}/tickets/{ticketId}/details` instead. Do not
       use this if the user's request implies they want to read discussion,
...
       conversation.
...

...

...
servers:
...
paths:
...
  /api/{provider}/containers:
    get:
      tags: [Discovery]
      operationId: getContainers
      summary: List the projects (Jira) or lists (ClickUp) tickets can be created in
      description: >-
        Use this route whenever you need a value for `projectKeyOrListId` to
        pass to `POST /api/{provider}/tickets`, or when the user asks "what
        projects/lists do we have" or "where should this ticket go". Every
        item's `id` field is exactly what you should pass back as
        `projectKeyOrListId`. If the user is asking about Scrum/Kanban
        boards specifically rather than where to file a ticket, use
        `GET /api/{provider}/boards` instead.


        For **Jira**, each item represents one project (`id` is the project
        key, e.g. `WEB`). For **ClickUp**, each item represents one list,
        flattened out of the team → space → folder → list hierarchy, with
        `path` showing that hierarchy for disambiguation (e.g. two lists
        named "Bugs" in different spaces will have different `path` values).


        This call can be slow for ClickUp (it walks every space and folder in
        the workspace) — avoid calling it more than once per conversation
        turn; cache the result if you need to look up multiple containers.
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
      responses:
        '200':
          description: List of containers (projects/lists) available to the configured credentials.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketContainer'
              examples:
                jira:
                  summary: Jira projects
                  value:
                    - id: WEB
                      name: MedEx.co
                    - id: MOB
                      name: MedEx Mobile
                clickup:
                  summary: ClickUp lists
                  value:
                    - id: '901307807729'
                      name: QuickBugFixes | Bipin
                      path: Current Core Dev / Projects
                    - id: '901313616559'
                      name: Bugs
                      path: Reseller Portal Bugs
        '400':
          $ref: '#/components/responses/BadProvider'
        '500':
          $ref: '#/components/responses/ProviderError'
  /api/{provider}/boards:
    get:
      tags: [Discovery]
      operationId: getBoards
      summary: List Scrum/Kanban boards (Jira) or board-viewable lists (ClickUp)
      description: >-
        Use this for "what boards do we have", "show me the Scrum/Kanban
        board for project X", or anything specifically about Agile
        board-style views — NOT for finding a `projectKeyOrListId` to create
        a ticket in (use `GET /api/{provider}/containers` for that instead;
        see the containers-vs-boards comparison in the top-level API
        description).


        For **Jira**, this calls the Agile REST API and returns real Scrum/
        Kanban board entities, each with a numeric `id` distinct from its
        project's key, plus `type` (`scrum`/`kanban`/`simple`) and
        `projectKeyOrListId` identifying which project it's attached to. A
        single project can have zero, one, or multiple boards, and a board's
        `id` is NOT usable as `projectKeyOrListId` when creating a ticket —
        use the `projectKeyOrListId` field on the board object for that, or
        just call the containers route directly.


        For **ClickUp**, there is no separate board entity — every list is
        viewable as a board — so this returns the same lists as the
        containers route, each tagged `type: "list"`, and its `id` IS
        directly usable as `projectKeyOrListId` (in ClickUp's case,
        `id` and `projectKeyOrListId` on each item are identical).


        Like the containers route, this can be slow for ClickUp since it
        walks the full space/folder hierarchy — avoid redundant calls within
        a conversation turn.
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
      responses:
        '200':
          description: Boards available to the configured credentials.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Board'
              examples:
                jira:
                  summary: Jira Agile boards
                  value:
                    - id: '38'
                      name: MDC board
                      type: simple
                      projectKeyOrListId: WEB
                    - id: '173'
                      name: OP board
                      type: scrum
                      projectKeyOrListId: OP
                clickup:
                  summary: ClickUp lists in board view
                  value:
                    - id: '901307807729'
                      name: QuickBugFixes | Bipin
                      type: list
                      projectKeyOrListId: '901307807729'
                      path: Current Core Dev / Projects
        '400':
          $ref: '#/components/responses/BadProvider'
        '500':
          $ref: '#/components/responses/ProviderError'
  /api/{provider}/users:
    get:
      tags: [Tickets]
      operationId: getUsers
      summary: Search assignable users
      description: >-
        Use to resolve a user's `id` before setting `assignee` on a
        create/update ticket request — never guess or hardcode an account
        ID. Optional `query` filters by name/email substring; omit it to
        list all users (ClickUp) or get an empty-query search (Jira).
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
        - name: query
          in: query
          required: false
          schema:
            type: string
          description: Name or email substring to filter users by.
      responses:
        '200':
          description: Users matching the query.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PMUser'
              examples:
                jira:
                  summary: Jira user search
                  value:
                    - id: '712020:a69b12f6-0f72-4a00-8283-ba9aa0fb01e5'
                      name: Nikesh Makaju
                      email: nikesh.m@medex.co.th
                clickup:
                  summary: ClickUp team member search
                  value:
                    - id: '118050764'
                      name: nikesh
                      email: n.makaju@a5sync.com
        '400':
          $ref: '#/components/responses/BadProvider'
        '500':
          $ref: '#/components/responses/ProviderError'
  /api/{provider}/tickets:
    get:
      tags: [Tickets]
      operationId: searchTickets
      summary: List/search tickets in a project or list, optionally filtered by status and/or assignee
      description: >-
        Use for "what tickets are in project X", "show open tickets in list
        Y", or "what's the most recent task assigned to user Z" — pass
        `assignee` (resolve the `id` via `GET /api/{provider}/users` first;
        never guess it) to filter to one person's tickets. Results are
        ordered most recently updated first, so the first item in the
        response is the most recent one.
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
        - name: projectKeyOrListId
          in: query
          required: true
          schema:
            type: string
          description: Jira project key or ClickUp list ID to search within.
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Filter to tickets in this status name.
        - name: assignee
          in: query
          required: false
          schema:
            type: string
          description: Jira account ID, or ClickUp numeric user ID (as a string). Resolve via `GET /api/{provider}/users`.
      responses:
        '200':
          description: Tickets matching the filters, most recently updated first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
        '400':
          description: Missing `projectKeyOrListId` query param, or unsupported provider.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ProviderError'

    post:
      tags: [Tickets]
      operationId: createTicket
      summary: Create a new ticket
      description: >-
        Use for "create a bug/task/issue/story" requests. Requires
        `projectKeyOrListId` — if the user hasn't specified which
        project/list, and there is any ambiguity, call
        `GET /api/{provider}/containers` first and confirm the target with the
        user rather than guessing. On success returns the full created ticket
        (equivalent shape to the "details" route, not the combined
        with-comments shape, since a brand-new ticket has no comments yet).
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketInput'
            examples:
              jira:
                summary: Create a Jira issue in project WEB
                value:
                  title: Fix checkout flow for logged in case
                  description: Guest checkout works; logged-in checkout is broken.
                  projectKeyOrListId: WEB
                  priority: High
              clickup:
                summary: Create a ClickUp task in a specific list
                value:
                  title: Fix checkout flow for logged in case
                  description: Guest checkout works; logged-in checkout is broken.
                  projectKeyOrListId: '901307807729'
                  priority: urgent
              jira-subtask:
                summary: Create a Jira sub-task under issue WEB-123
                value:
                  title: Write regression test for checkout fix
                  description: Cover the logged-in checkout path.
                  projectKeyOrListId: WEB
                  parentId: WEB-123
              clickup-subtask:
                summary: Create a ClickUp subtask under an existing task
                value:
                  title: Write regression test for checkout fix
                  description: Cover the logged-in checkout path.
                  projectKeyOrListId: '901307807729'
                  parentId: '86b3xyz123'
      responses:
        '201':
          description: Ticket created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          description: >-
            Missing/invalid `projectKeyOrListId`, unsupported provider, or the
            underlying tool rejected the payload (e.g. unknown project key,
            unknown list ID, or invalid priority value for that provider).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ProviderError'
  /api/{provider}/tickets/assigned:
    get:
      tags: [Tickets]
      operationId: searchTicketsAssignedToUser
      summary: List tickets assigned to a user across the entire workspace (not scoped to one project/list)
      description: >-
        Use for "list all tasks assigned to <person>" or "what's on my plate"
        when the user hasn't named a specific project/list — unlike
        `GET /api/{provider}/tickets`, this searches every project/list at
        once. Pass `assignee` (resolve the `id` via
        `GET /api/{provider}/users` first; never guess it). For **ClickUp**,
        `containerId` (the workspace/team ID from
        `GET /api/{provider}/workspaces`) is REQUIRED, since ClickUp's
        cross-list task search is workspace-scoped. Jira ignores
        `containerId` since JQL can search across all projects directly.
        Results are ordered most recently updated first.
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
        - name: assignee
          in: query
          required: true
          schema:
            type: string
          description: Jira account ID, or ClickUp numeric user ID (as a string). Resolve via `GET /api/{provider}/users`.
        - name: containerId
          in: query
          required: false
          schema:
            type: string
          description: Required for ClickUp (workspace/team ID from `GET /api/{provider}/workspaces`). Not used by Jira.
      responses:
        '200':
          description: Tickets assigned to the user, most recently updated first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
        '400':
          description: Missing `assignee` query param (or missing `containerId` for ClickUp), or unsupported provider.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ProviderError'
  /api/{provider}/tickets/{ticketId}:
    get:
      tags: [Tickets]
      operationId: getTicketWithComments
      summary: Get a ticket AND its comments in one call (default "show me this ticket" route)
      description: >-
        This is the route to use for the general case of "show me ticket X",
        "what's the status of X", "summarize this ticket", "what's the
        discussion on X" — anything where the user wants to see or reason
        about a ticket. It returns everything `GET .../details` returns, plus
        a `comments` array, in a single round trip. Prefer this over calling
        `.../details` and `.../comments` separately unless you have a specific
        reason to fetch only one of the two (e.g. performance-sensitive bulk
        fetch of many tickets where comments aren't needed — use `.../details`
        for that case instead).
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
        - $ref: '#/components/parameters/TicketIdParam'
      responses:
        '200':
          description: Ticket with its comments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketWithComments'
        '400':
          $ref: '#/components/responses/BadProvider'
        '404':
          $ref: '#/components/responses/TicketNotFound'
        '500':
          $ref: '#/components/responses/ProviderError'

    patch:
      tags: [Tickets]
      operationId: updateTicket
      summary: Update ticket fields (title, description, assignee, priority)
      description: >-
        Use for "rename this ticket", "update the description", "reassign
        this to X", "change priority to Y". All fields are optional — only
        send the ones you want to change; omitted fields are left untouched.

        Do NOT use this route to change ticket status/workflow state — send
        only field edits here. For status changes use
        `PATCH /api/{provider}/tickets/{ticketId}/status` instead, since
        status is transition-gated (especially in Jira) and handled
        separately from arbitrary field edits.
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
        - $ref: '#/components/parameters/TicketIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTicketInput'
            examples:
              reassignAndReprioritize:
                summary: Change assignee and priority only
                value:
                  assignee: jane.doe
                  priority: High
      responses:
        '200':
          description: Updated ticket (fields only — no comments; call the GET route if you need those too).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          $ref: '#/components/responses/BadProvider'
        '404':
          $ref: '#/components/responses/TicketNotFound'
        '500':
          $ref: '#/components/responses/ProviderError'

    delete:
      tags: [Tickets]
      operationId: deleteTicket
      summary: Permanently delete a ticket
      description: >-
        DESTRUCTIVE AND IRREVERSIBLE in both Jira and ClickUp — there is no
        "undo" or trash/recycle-bin step exposed by this API. Only call this
        when the user has explicitly and specifically asked to delete a named
        ticket by ID/key. Never call this to "clean up" or as a guess when a
        create/update fails. If there's any doubt about which ticket the user
        means, confirm the exact ticket ID/key before calling this route.
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
        - $ref: '#/components/parameters/TicketIdParam'
      responses:
        '204':
          description: Ticket deleted. No response body.
        '400':
          $ref: '#/components/responses/BadProvider'
        '404':
          $ref: '#/components/responses/TicketNotFound'
        '500':
          $ref: '#/components/responses/ProviderError'
  /api/{provider}/tickets/{ticketId}/details:
    get:
      tags: [Tickets]
      operationId: getTicket
      summary: Get ticket fields only, WITHOUT comments (lighter/faster than the combined route)
      description: >-
        Returns the same ticket fields as `GET /api/{provider}/tickets/{ticketId}`
        but skips fetching comments — use this when you explicitly don't need
        discussion/comments (e.g. bulk-checking status/assignee across many
        tickets, or when you know the user only cares about a specific field
        like status or priority). If there is any chance the user wants to
        see discussion/history, use the combined
        `GET /api/{provider}/tickets/{ticketId}` route instead — don't use
        this route as your default.
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
        - $ref: '#/components/parameters/TicketIdParam'
      responses:
        '200':
          description: Ticket fields, no comments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          $ref: '#/components/responses/BadProvider'
        '404':
          $ref: '#/components/responses/TicketNotFound'
        '500':
          $ref: '#/components/responses/ProviderError'
  /api/{provider}/tickets/{ticketId}/comments:
    get:
      tags: [Comments]
      operationId: getComments
      summary: List a ticket's comments only (no ticket fields)
      description: >-
        Use only when you specifically need just the discussion/comment
        thread and not the ticket's own fields (title/status/etc.) — e.g.
        "summarize the conversation on this ticket" without also needing its
        metadata. If you need both, prefer the combined
        `GET /api/{provider}/tickets/{ticketId}` route to save a round trip.
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
        - $ref: '#/components/parameters/TicketIdParam'
      responses:
        '200':
          description: List of comments, oldest first as returned by the underlying provider.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketComment'
        '400':
          $ref: '#/components/responses/BadProvider'
        '404':
          $ref: '#/components/responses/TicketNotFound'
        '500':
          $ref: '#/components/responses/ProviderError'

    post:
      tags: [Comments]
      operationId: addComment
      summary: Add a comment to a ticket
      description: >-
        Use for "add a comment/note/reply to this ticket", "post an update on
        X". This only adds to the discussion thread — it does not change the
        ticket's status or any other field. To also change status, make a
        separate call to the status route.
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
        - $ref: '#/components/parameters/TicketIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentInput'
            examples:
              default:
                value:
                  body: Verified the fix in staging — looks good to ship.
      responses:
        '201':
          description: Comment created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketComment'
        '400':
          $ref: '#/components/responses/BadProvider'
        '404':
          $ref: '#/components/responses/TicketNotFound'
        '500':
          $ref: '#/components/responses/ProviderError'
  /api/{provider}/tickets/{ticketId}/statuses:
    get:
      tags: [Status]
      operationId: getAvailableStatuses
      summary: List the statuses this specific ticket can currently move to
      description: >-
        ALWAYS call this before `PATCH /api/{provider}/tickets/{ticketId}/status`
        if you are not 100% certain the target status name/id is valid for
        this exact ticket. The result is ticket-specific, not just
        provider-wide:

        - In **Jira**, this returns the workflow *transitions* currently
          available from the ticket's current status — not every status in
          the project's workflow is reachable from every other status, so the
          same issue type in the same project can have different available
          transitions depending on its current state.
        - In **ClickUp**, this returns every status defined on the ticket's
          list (ClickUp lets you jump to any status directly, no transition
          graph), so the result only depends on which list the task is in.

        Each returned item's `id` or `name` can be used as the `status` value
        in the status-update request body.
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
        - $ref: '#/components/parameters/TicketIdParam'
      responses:
        '200':
          description: Statuses/transitions available for this ticket right now.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketStatus'
              examples:
                jira:
                  summary: Available Jira transitions from "In Progress"
                  value:
                    - id: '4'
                      name: IN QA
                    - id: '2'
                      name: CLOSE
                clickup:
                  summary: All statuses on the ClickUp task's list
                  value:
                    - id: p90132474781_DCnaeiSB
                      name: to do
                      category: open
                    - id: p90132474781_g9uxhsQM
                      name: in progress
                      category: custom
        '400':
          $ref: '#/components/responses/BadProvider'
        '404':
          $ref: '#/components/responses/TicketNotFound'
        '500':
          $ref: '#/components/responses/ProviderError'
...
components:
  parameters:
    ProviderParam:
      name: provider
      in: path
      required: true
      description: >-
        Which PM tool to talk to. Must be exactly `jira` or `clickup` — any
        other value returns a 400 before any provider call is made.
      schema:
        type: string
        enum: [jira, clickup]
      example: jira
...
  responses:
    BadProvider:
      description: >-
        `provider` path segment was not `jira` or `clickup`. Returned before
        any request reaches the underlying tool.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'Unsupported provider "foo". Supported providers: jira, clickup'
...
  schemas:
...
    TicketWithComments:
      allOf:
        - $ref: '#/components/schemas/Ticket'
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: '#/components/schemas/TicketComment'
          required: [comments]
...